+++ sciris/sc_math.py
@@ -281,35 +281,31 @@ def getvaliddata(data=None, filterdata=None, defaultind=0): # pragma: no cover
     return validdata
 
 
-def sanitize(data=None, returninds=False, replacenans=None, die=True, defaultval=None, label=None, verbose=True):
+def sanitize(data=None, returninds=False, replacenans=None, defaultval=None, die=True, verbose=False, label=None):
         '''
-        Sanitize input to remove NaNs. Warning, does not work on multidimensional data!!
-        Returns an array with the sanitized data. If replacenans=True, the sanitized array is
-        of the same length/size as data. If replacenans=False, the sanitized array
-        may be shorter than data.
+        Sanitize input to remove NaNs.
+
+        Returns an array with the sanitized data. If ``replacenans=True``, the sanitized
+        array is of the same length/size as data. If ``replacenans=False``, the sanitized
+        array may be shorter than data.
 
         Args:
-            data        (arr/list)   : the array with data to be sanitized
-            returninds  (bool)       : whether to return the indices of the non-NaN values in data
-            replacenans (bool/float) : whether to replace the NaNs with a value, or using interpolation ``sc.smoothinterp()``
-            defaultval  (float/int)  : value to return if the sanitized array is empty.
-            label       (str)        : human readable label for data
-            die         (bool)       : whether to raise an exception if sanitization fails.
-
-        Args:
-        data (array): array or list with numbers
-        returninds (bool): whether to return indices of non-nan/valid  elements, indices are with respect the shape of data
-        die (bool): whether to raise an exception if the sanitization failed (otherwise return an empty array)
-        defaultval (float): if not valid data are found in data, return defaultval. If defaultval=None (default), returns 0
-        label (str): human readable label to describe data
-        verbose (bool): whether to report progress
+            data        (arr/list)   : array or list with numbers to be sanitized
+            returninds  (bool)       : whether to return indices of non-nan/valid elements, indices are with respect the shape of data
+            replacenans (float/str)  : whether to replace the NaNs with the specified value, or if ``True`` or a string, using interpolation
+            defaultval  (float)      : value to return if the sanitized array is empty
+            die         (bool)       : whether to raise an exception if the sanitization failed (otherwise return an empty array)
+            verbose     (bool)       : whether to print out a warning if no valid values are found
+            label       (str)        : human readable label for data (for use with verbose mode only)
 
         **Examples**::
 
             data = [3, 4, np.nan, 8, 2, np.nan, np.nan, 8]
-            sanitized1, inds = sc.sanitize(data, returninds=True)
-            sanitized2 = sc.sanitize(data, replacenans=True)
-            sanitized3 = sc.sanitize(data, replacenans=0)
+            sanitized1, inds = sc.sanitize(data, returninds=True) # Remove NaNs
+            sanitized2 = sc.sanitize(data, replacenans=True) # Replace NaNs using nearest neighbor interpolation
+            sanitized3 = sc.sanitize(data, replacenans='nearest') # Eequivalent to replacenans=True
+            sanitized4 = sc.sanitize(data, replacenans='linear') # Replace NaNs using linear interpolation
+            sanitized5 = sc.sanitize(data, replacenans=0) # Replace NaNs with 0
         '''
         try:
             data = np.array(data, dtype=float) # Make sure it's an array of float type
@@ -320,8 +316,12 @@ def sanitize(data=None, returninds=False, replacenans=None, die=True, defaultval
             if replacenans is not None:
                 newx = range(len(data)) # Create a new x array the size of the original array
                 if replacenans==True: replacenans = 'nearest'
-                if replacenans in ['nearest','linear']:
-                    sanitized = smoothinterp(newx, inds, sanitized, method=replacenans, smoothness=0) # Replace nans with interpolated values
+                if scu.isstring(replacenans):
+                    if replacenans in ['nearest','linear']:
+                        sanitized = smoothinterp(newx, inds, sanitized, method=replacenans, smoothness=0) # Replace nans with interpolated values
+                    else:
+                        errormsg = f'Interpolation method "{replacenans}" not found: must be "nearest" or "linear"'
+                        raise ValueError(errormsg)
                 else:
                     naninds = inds = np.nonzero(np.isnan(data))[0]
                     sanitized = scu.dcp(data)
@@ -332,7 +332,7 @@ def sanitize(data=None, returninds=False, replacenans=None, die=True, defaultval
                 else:
                     sanitized = data
                     inds = []
-                    
+
                     if verbose: # pragma: no cover
                         if label is None: label = 'these input data'
                         print(f'sc.sanitize(): no valid values found for {label}. Returning 0.')
@@ -349,16 +349,19 @@ def sanitize(data=None, returninds=False, replacenans=None, die=True, defaultval
 
 def rmnans(*args, **kwargs):
     """
-    Alias for sanitize(...) with default values. New in version 1.4.0. '''
+    Alias for ``sc.sanitize()``
     """
     return sanitize(*args, **kwargs)
+rmnans.__doc__ += '\n\n' + sanitize.__doc__
 
 
-def fillnans(*args, **kwargs):
+def fillnans(data=None, *args, replacenans=True, **kwargs):
     """
-    Alias for sanitize(..., replacenans=True) with nearest interpolatino. New in version 1.4.0. '''
+    Alias for ``sc.sanitize(..., replacenans=True) with nearest interpolation (or a specified  value).
     """
     return sanitize(*args, **kwargs, replacenans=True)
+fillnans.__doc__ += '\n\n' + sanitize.__doc__
+
 
 def isprime(n, verbose=False):
     '''