+++ sciris/sc_utils.py
@@ -43,7 +43,7 @@ from distutils.version import LooseVersion
 
 # Handle types
 _stringtypes = (str, bytes)
-_numtype    = numbers.Number
+_numtype     = numbers.Number
 
 
 ##############################################################################
@@ -64,10 +64,12 @@ def fast_uuid(which=None, length=None, n=1, secure=False, forcelist=False, safet
         which (str): the set of characters to choose from (default ascii)
         length (int): length of UID (default 6)
         n (int): number of UIDs to generate
+        secure (bool): whether to generate random numbers from sources provided by the operating system
         forcelist (bool): whether or not to return a list even for a single UID (used for recursive calls)
         safety (float): ensure that the space of possible UIDs is at least this much larger than the number requested
         recursion (int): the recursion level of the call (since the function calls itself if not all UIDs are unique)
-        recursion_limit (int): # Maximum number of times to try regeneraring keys
+        recursion_limit (int): Maximum number of times to try regeneraring keys
+        verbose (bool): whether to show progress
 
     Returns:
         uid (str or list): a string UID, or a list of string UIDs
@@ -228,35 +230,42 @@ def uuid(uid=None, which=None, die=False, tostring=False, length=None, n=1, **kw
     return output
 
 
-def dcp(obj, verbose=True, die=False):
+def dcp(obj, die=True, verbose=True):
     '''
     Shortcut to perform a deep copy operation
 
-    Almost identical to ``copy.deepcopy()``
+    Almost identical to ``copy.deepcopy()``, but optionally fall back to copy()
+    if deepcopy fails.
+
+    Args:
+        die (bool): if False, fall back to copy()
+        verbose (bool): if die is False, then print a warning if deepcopy() fails
+
+    New in version 1.4.0: default die=True instead of False
     '''
     try:
         output = copy.deepcopy(obj)
     except Exception as E: # pragma: no cover
         output = cp(obj)
-        errormsg = f'Warning: could not perform deep copy, performing shallow instead: {str(E)}'
+        errormsg = f'Warning: could not perform deep copy: {str(E)}'
         if die: raise RuntimeError(errormsg)
-        else:   print(errormsg)
+        else:   print(errormsg + '\nPerforming shallow copy instead...')
     return output
 
 
-def cp(obj, verbose=True, die=True):
+def cp(obj, die=True, verbose=True):
     '''
     Shortcut to perform a shallow copy operation
 
-    Almost identical to ``copy.copy()``
+    Almost identical to ``copy.copy()``, but optionally allow failures
     '''
     try:
         output = copy.copy(obj)
     except Exception as E:
         output = obj
-        errormsg = 'Could not perform shallow copy, returning original object'
+        errormsg = 'Could not perform shallow copy'
         if die: raise ValueError(errormsg) from E
-        else:   print(errormsg)
+        else:   print(errormsg + '\nReturning original object...')
     return output
 
 
@@ -467,12 +476,17 @@ def require(reqs=None, *args, exact=False, detailed=False, die=True, verbose=Tru
 
     # Handle exceptions
     if not met:
-        errormsg = 'The following requirements were not met:'
+        errkeys = list(errs.keys())
+        errormsg = '\nThe following requirement(s) were not met:'
+        count = 0
         for k,v in data.items():
             if not v:
-                errormsg += f'\n  {k}: {str(errs[k])}'
+                count += 1
+                errormsg += f'\n• "{k}": {str(errs[k])}'
+        errormsg += f'''\n\nIf this is a valid module, you might want to try "pip install {strjoin(errkeys, sep=' ')} --upgrade".'''
         if die:
-            raise ModuleNotFoundError(errormsg) from errs[k] # Use the last one
+            err = errs[errkeys[-1]]
+            raise ModuleNotFoundError(errormsg) from err
         elif verbose:
             print(errormsg)
 
@@ -702,13 +716,14 @@ def isarray(obj, dtype=None):
 def promotetoarray(x, keepnone=False, **kwargs):
     '''
     Small function to ensure consistent format for things that should be arrays
-    (note: toarray()/promotetoarray() are identical).
+    (note: ``toarray()`` and ``promotetoarray()`` are identical).
 
     Very similar to ``np.array``, with the main difference being that ``sc.promotetoarray(3)``
     will return ``np.array([3])`` (i.e. a 1-d array that can be iterated over), while
     ``np.array(3)`` will return a 0-d array that can't be iterated over.
 
     Args:
+        x (any): a number or list of numbers
         keepnone (bool): whether ``sc.promotetoarray(None)`` should return ``np.array([])`` or ``np.array([None], dtype=object)``
         kwargs (dict): passed to ``np.array()``
 
@@ -725,7 +740,7 @@ def promotetoarray(x, keepnone=False, **kwargs):
     if skipnone is not None: # pragma: no cover
         keepnone = not(skipnone)
         warnmsg = 'sc.promotetoarray() argument "skipnone" has been deprecated as of v1.1.0; use keepnone instead'
-        warnings.warn(warnmsg, category=DeprecationWarning, stacklevel=2)
+        warnings.warn(warnmsg, category=FutureWarning, stacklevel=2)
     if isnumber(x) or (isinstance(x, np.ndarray) and not np.shape(x)): # e.g. 3 or np.array(3)
         x = [x]
     elif x is None and not keepnone:
@@ -901,8 +916,8 @@ def mergedicts(*args, _strict=False, _overwrite=True, _copy=False, **kwargs):
     # Warn about deprecated keys
     renamed = ['strict', 'overwrite', 'copy']
     if any([k in kwargs for k in renamed]):
-            warnmsg = f'sc.mergedicts() arguments "{strjoin(renamed)}" have been renamed with underscores as of v1.3.3; using these as keywords is undesirable'
-            warnings.warn(warnmsg, category=DeprecationWarning, stacklevel=2)
+        warnmsg = f'sc.mergedicts() arguments "{strjoin(renamed)}" have been renamed with underscores as of v1.3.3; using these as keywords is undesirable'
+        warnings.warn(warnmsg, category=FutureWarning, stacklevel=2)
 
     # Try to get the output type from the first argument, but revert to a standard dict if that fails
     try:
@@ -998,7 +1013,7 @@ def _sanitize_output(obj, is_list, is_array, dtype=None):
 #%% Misc. functions
 ##############################################################################
 
-__all__ += ['strjoin', 'newlinejoin', 'checkmem', 'checkram', 'runcommand', 'gitinfo',
+__all__ += ['strjoin', 'newlinejoin', 'strsplit', 'checkmem', 'checkram', 'runcommand', 'gitinfo',
             'compareversions', 'uniquename', 'importbyname', 'suggest', 'profile', 'mprofile',
             'getcaller']
 
@@ -1041,6 +1056,53 @@ def newlinejoin(*args):
     New in version 1.1.0.
     '''
     return strjoin(*args, sep='\n')
+
+
+def strsplit(string, sep=None, skipempty=True, lstrip=True, rstrip=True):
+    '''
+    Convenience function to split common types of strings.
+
+    Note: to use regular expressions, use ``re.split()`` instead.
+
+    Args:
+        string    (str):      the string to split
+        sep       (str/list): the types of separator to accept (default space or comma, i.e. [' ', ','])
+        skipempty (bool):     whether to skip empty entries (i.e. from consecutive delimiters)
+        lstrip    (bool):     whether to strip any extra spaces on the left
+        rstrip    (bool):     whether to strip any extra spaces on the right
+
+    Examples:
+
+        sc.strsplit('a b c') # Returns ['a', 'b', 'c']
+        sc.strsplit('a,b,c') # Returns ['a', 'b', 'c']
+        sc.strsplit('a, b, c') # Returns ['a', 'b', 'c']
+        sc.strsplit('  foo_bar  ', sep='_') # Returns ['foo', 'bar']
+
+    New in version 2.0.0.
+    '''
+    strlist = []
+    if sep is None:
+        sep = [' ', ',']
+
+    # Generate a character sequence that isn't in the string
+    special = '∙' # Pick an obscure character
+    while special in string: # If it exists in the string nonetheless, keep going until it doesn't
+        special += special
+
+    # Convert all separators to the special character
+    for s in sep:
+        string = string.replace(s, special)
+
+    # Split the string, filter, and trim
+    strlist = string.split(special)
+    if lstrip:
+        strlist = [s.lstrip() for s in strlist]
+    if rstrip:
+        strlist = [s.rstrip() for s in strlist]
+    if skipempty:
+        strlist = [s for s in strlist if s != '']
+
+    return strlist
 
 
 def checkmem(var, descend=None, alphabetical=False, plot=False, verbose=False):
@@ -1419,6 +1481,7 @@ def suggest(user_input, valid_inputs, n=1, threshold=None, fulloutput=False, die
         valid_inputs (list): List/collection of valid strings
         n (int): Maximum number of suggestions to return
         threshold (int): Maximum number of edits required for an option to be suggested (by default, two-thirds the length of the input; for no threshold, set to -1)
+        fulloutput (bool): Whether to return suggestions and distances.
         die (bool): If True, an informative error will be raised (to avoid having to implement this in the calling code)
         which (str): Distance calculation method used; options are "damerau" (default), "levenshtein", or "jaro"